CreateML/requirements.txt
# Core ML and Data Processing
torch>=2.0.0
torchvision>=0.15.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# NLP Features
sentence-transformers>=2.2.0
textblob>=0.17.0
nltk>=3.8.0

# Gemini Features (CLIP + VLLM)
transformers>=4.35.0
accelerate>=0.25.0
anthropic>=0.39.0

# Utilities
pillow>=10.0.0
tqdm>=4.65.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
